Re-org conftest setup
